Remove redundant UI components and greated a general data display UI component for futrue data display screens
