+++ src/main/java/seedu/address/ui/DataPanel.java
@@ -1,0 +1,9 @@
+package seedu.address.ui;
+
+import javafx.scene.layout.Region;
+
+public class DataPanel extends UiPart<Region> {
+    public DataPanel(String name) {
+        super(name);
+    }
+}
