--- src/main/java/seedu/address/ui/StudentListPanel.java
@@ -1,47 +1,0 @@
-package seedu.address.ui;
-
-import java.util.logging.Logger;
-
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.scene.control.ListCell;
-import javafx.scene.control.ListView;
-import javafx.scene.layout.Region;
-import seedu.address.commons.core.LogsCenter;
-import seedu.address.model.group.Student;
-
-public class StudentListPanel extends UiPart<Region> {
-
-    private static final String FXML = "StudentListPanel.fxml";
-    private final Logger logger = LogsCenter.getLogger(PersonListPanel.class);
-
-    @FXML
-    private ListView<Student> studentListView;
-
-    /**
-     * Creates a {@code PersonListPanel} with the given {@code ObservableList}.
-     */
-    public StudentListPanel(ObservableList<Student> studentList) {
-        super(FXML);
-        studentListView.setItems(studentList);
-        studentListView.setCellFactory(listView -> new StudentListViewCell());
-    }
-
-    /**
-     * Custom {@code ListCell} that displays the graphics of a {@code Student} using a {@code StudentCard}.
-     */
-    class StudentListViewCell extends ListCell<Student> {
-
-        @Override
-        protected void updateItem(Student student, boolean empty) {
-            super.updateItem(student, empty);
-
-            if (empty || student == null) {
-                setGraphic(null);
-                setText(null);
-            } else {
-                setGraphic(new StudentCard(student, getIndex() + 1).getRoot());
-            }
-        }
-    }
-}
