+++ src/main/resources/view/GroupDataPanel.fxml
@@ -1,21 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.layout.*?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.control.SplitPane?>
+<?import javafx.scene.layout.VBox?>
 <?import javafx.geometry.Insets?>
-
-<HBox alignment="CENTER" prefHeight="100.0" prefWidth="200.0" VBox.vgrow="ALWAYS" xmlns:fx="http://javafx.com/fxml">
-    <children>
-        <VBox fx:id="studentList" alignment="CENTER" minWidth="340" prefWidth="340" styleClass="pane-with-border" HBox.hgrow="ALWAYS">
-            <padding>
-                <Insets bottom="10" left="10" right="10" top="10" />
-            </padding>
-            <StackPane fx:id="studentListPanelPlaceholder" VBox.vgrow="ALWAYS" />
-        </VBox>
-        <VBox fx:id="lessonList" alignment="CENTER" minWidth="340" prefWidth="340" styleClass="pane-with-border" HBox.hgrow="ALWAYS">
-            <padding>
-                <Insets bottom="10" left="10" right="10" top="10" />
-            </padding>
-            <StackPane fx:id="lessonListPanelPlaceholder" VBox.vgrow="ALWAYS" />
-        </VBox>
-    </children>
-</HBox>
+<SplitPane dividerPositions="0.5" minWidth="680.0" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
+   <padding>
+      <Insets bottom="5" left="10" right="10" top="5" />
+   </padding>
+   <items>
+      <VBox prefHeight="200.0" prefWidth="100.0">
+         <ListView fx:id="studentListView" VBox.vgrow="ALWAYS" />
+      </VBox>
+      <VBox prefHeight="200.0" prefWidth="100.0">
+         <ListView fx:id="lessonListView" VBox.vgrow="ALWAYS" />
+      </VBox>
+   </items>
+</SplitPane>
