--- src/main/resources/view/StudentInfoListPanel.fxml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.layout.VBox?>
-
-<VBox xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml">
-    <ListView fx:id="studentInfoListView" VBox.vgrow="ALWAYS" />
-</VBox>
